# backend/src/main/resources/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlayerSignal local alpha",
    "version": "0.6.0",
    "description": "Accounts are opt-in via AUTH_ENABLED. When enabled, game resources require a session and every mutation requires X-CSRF-TOKEN from /api/auth/session. Cross-workspace resources return 404."
  },
  "paths": {
...
    "/api/games/{game}/issues": {
      "parameters": [
        {
          "name": "game",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Issue snapshot ordered by mean severity then mention count",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "minSeverity",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePage"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/games/{game}/issues/rebuild": {
      "parameters": [
        {
          "name": "game",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "summary": "Atomically replace clusters from current actionable analyses; local lexical vectors",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePage"
                }
              }
            }
          },
          "409": {
            "description": "Rebuild already running"
          },
          "422": {
            "description": "More than 2000 eligible analyses; previous snapshot preserved"
          },
          "503": {
            "description": "Rebuild failed; previous snapshot preserved"
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required when accounts are enabled.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/games/{game}/issues/{issue}": {
      "parameters": [
        {
          "name": "game",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "issue",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Issue metrics and paginated exact source evidence",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetail"
                }
              }
            }
          },
          "404": {
            "description": "Issue not found in selected game"
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
    "/api/games/{id}/overview": {
      "get": {
        "summary": "Live overview with equal-duration comparison windows and UTC daily buckets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "enum": [
                7,
                30,
                90
              ],
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overview snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported range"
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/overview/demo": {
      "get": {
        "summary": "Ten-review fixed synthetic dashboard; read-only and no live calls",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "enum": [
                7,
                30,
                90
              ],
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overview snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported range"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "required": [
          "id",
          "steamAppId",
          "name",
          "headerImageUrl",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "steamAppId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "headerImageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Run": {
        "type": "object",
        "required": [
          "id",
          "gameId",
          "status",
          "startedAt",
          "finishedAt",
          "fetched",
          "inserted",
          "updated",
          "nextCursor",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "FAILED"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fetched": {
            "type": "integer",
            "minimum": 0
          },
          "inserted": {
            "type": "integer",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Classification": {
        "type": "object",
        "properties": {
          "sentiment": {
            "type": "string",
            "enum": [
              "VERY_NEGATIVE",
              "NEGATIVE",
              "MIXED",
              "POSITIVE",
              "VERY_POSITIVE"
            ]
          },
          "primaryCategory": {
            "type": "string",
            "enum": [
              "BUG",
              "PERFORMANCE",
              "GAMEPLAY",
              "UI_UX",
              "MULTIPLAYER",
              "BALANCE",
              "CONTENT",
              "CONTROLS",
              "AUDIO",
              "LOCALIZATION",
              "POSITIVE",
              "OTHER"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "normalizedIssue": {
            "type": "string",
            "maxLength": 240
          },
          "isActionable": {
            "type": "boolean"
          },
          "isLikelyBug": {
            "type": "boolean"
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 500
            },
            "maxItems": 5
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 40
            },
            "maxItems": 8
          }
        },
        "required": [
          "sentiment",
          "primaryCategory",
          "severity",
          "confidence",
          "normalizedIssue",
          "isActionable",
          "isLikelyBug",
          "evidence",
          "tags"
        ],
        "additionalProperties": false
      },
...
      "AnalysisRun": {
        "type": "object",
        "required": [
          "id",
          "gameId",
          "provider",
          "model",
          "promptVersion",
          "startedAt",
          "status",
          "processed",
          "succeeded",
          "skipped",
          "failed",
          "cached",
          "finishedAt",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "gameId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "startedAt": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ]
          },
          "processed": {
            "type": "integer",
            "minimum": 0
          },
          "succeeded": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "minimum": 0
          },
          "cached": {
            "type": "integer",
            "minimum": 0
          },
          "finishedAt": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AnalysisStatus": {
        "type": "object",
        "required": [
          "available",
          "provider",
          "model",
          "promptVersion",
          "maxReviewsPerRun",
          "counts",
          "latestRun"
        ],
        "properties": {
          "available": {
            "type": "boolean"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "maxReviewsPerRun": {
            "type": "integer",
            "minimum": 0
          },
          "counts": {
            "type": "object",
            "required": [
              "total",
              "pending",
              "running",
              "succeeded",
              "skipped",
              "failed"
            ],
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0
              },
              "pending": {
                "type": "integer",
                "minimum": 0
              },
              "running": {
                "type": "integer",
                "minimum": 0
              },
              "succeeded": {
                "type": "integer",
                "minimum": 0
              },
              "skipped": {
                "type": "integer",
                "minimum": 0
              },
              "failed": {
                "type": "integer",
                "minimum": 0
              }
            }
          },
          "latestRun": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AnalysisRun"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "IssueMetrics": {
        "type": "object",
        "required": [
          "mentionCount",
          "recentMentions",
          "previousMentions",
          "firstSeenAt",
          "lastSeenAt",
          "negativeRatio",
          "severityScore",
          "confidence",
          "velocityPercent"
        ],
        "properties": {
          "mentionCount": {
            "type": "integer",
            "minimum": 0
          },
          "recentMentions": {
            "type": "integer",
            "minimum": 0
          },
          "previousMentions": {
            "type": "integer",
            "minimum": 0
          },
          "firstSeenAt": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string"
          },
          "negativeRatio": {
            "type": "number"
          },
          "severityScore": {
            "type": "number"
          },
          "confidence": {
            "type": "number"
          },
          "velocityPercent": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "Issue": {
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "status",
          "metrics"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/IssueMetrics"
          }
        }
      },
      "IssueSnapshot": {
        "type": "object",
        "required": [
          "builtAt",
          "algorithm",
          "threshold",
          "clusteredReviews",
          "eligibleReviews",
          "stale"
        ],
        "properties": {
          "builtAt": {
            "type": "string",
            "nullable": true
          },
          "algorithm": {
            "type": "string"
          },
          "threshold": {
            "type": "number"
          },
          "clusteredReviews": {
            "type": "integer",
            "minimum": 0
          },
          "eligibleReviews": {
            "type": "integer",
            "minimum": 0
          },
          "stale": {
            "type": "boolean"
          }
        }
      },
      "IssueEvidence": {
        "type": "object",
        "required": [
          "analysisId",
          "reviewId",
          "steamRecommendationId",
          "sourceText",
          "language",
          "seenAt",
          "model",
          "promptVersion",
          "classification",
          "similarity"
        ],
        "properties": {
          "analysisId": {
            "type": "string"
          },
          "reviewId": {
            "type": "string"
          },
          "steamRecommendationId": {
            "type": "string"
          },
          "sourceText": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "seenAt": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "similarity": {
            "type": "number"
          }
        }
      },
      "IssuePage": {
        "type": "object",
        "required": [
          "snapshot",
          "items",
          "page",
          "size",
          "total"
        ],
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/IssueSnapshot"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "IssueDetail": {
        "type": "object",
        "required": [
          "snapshot",
          "issue",
          "evidence",
          "page",
          "size",
          "total"
        ],
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/IssueSnapshot"
          },
          "issue": {
            "$ref": "#/components/schemas/Issue"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueEvidence"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "OverviewMetrics": {
        "type": "object",
        "required": [
          "reviews",
          "recommended",
          "previousReviews",
          "positiveRatio",
          "previousPositiveRatio",
          "reviewChangePercent",
          "positiveChangePoints"
        ],
        "properties": {
          "reviews": {
            "type": "integer",
            "minimum": 0
          },
          "recommended": {
            "type": "integer",
            "minimum": 0
          },
          "previousReviews": {
            "type": "integer",
            "minimum": 0
          },
          "positiveRatio": {
            "type": "number",
            "nullable": true
          },
          "previousPositiveRatio": {
            "type": "number",
            "nullable": true
          },
          "reviewChangePercent": {
            "type": "number",
            "nullable": true
          },
          "positiveChangePoints": {
            "type": "number",
            "nullable": true
          }
        }
      },
      "OverviewDay": {
        "type": "object",
        "required": [
          "date",
          "recommended",
          "notRecommended"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "recommended": {
            "type": "integer",
            "minimum": 0
          },
          "notRecommended": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "OverviewCategory": {
        "type": "object",
        "required": [
          "category",
          "count"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Overview": {
        "type": "object",
        "required": [
          "mode",
          "game",
          "days",
          "periodStart",
          "asOf",
          "imported",
          "metrics",
          "daily",
          "categories",
          "analysis",
          "ingestion",
          "issues",
          "highSeverityIssues"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "LIVE",
              "SYNTHETIC_DEMO"
            ]
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "days": {
            "type": "integer",
            "enum": [
              7,
              30,
              90
            ]
          },
          "periodStart": {
            "type": "string"
          },
          "asOf": {
            "type": "string"
          },
          "imported": {
            "type": "integer",
            "minimum": 0
          },
          "metrics": {
            "$ref": "#/components/schemas/OverviewMetrics"
          },
          "daily": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OverviewDay"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OverviewCategory"
            }
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "ingestion": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Run"
              }
            ],
            "nullable": true
          },
          "issues": {
            "$ref": "#/components/schemas/IssuePage"
          },
          "highSeverityIssues": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "PLAYERSIGNAL_SESSION"
      }
    }
  }
}